backend/requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-learn
numpy
passlib[bcrypt]
python-jose[cryptography]
python-multipart